feat: add `GradCkptConfig` and `GradCkptCollection`
