+++ colossalai/shardformer/__init__.py
@@ -1,1 +1,1 @@
-from .shard import PipelineGradientConfig, ShardConfig, ShardFormer
+from .shard import GradCkptCollection, ModelSharder, PipelineGradCkptConfig, ShardConfig, ShardFormer
