+++ colossalai/shardformer/shard/__init__.py
@@ -1,5 +1,6 @@
-from .shard_config import PipelineGradientConfig, ShardConfig
+from .grad_ckpt_config import GradCkptCollection, PipelineGradCkptConfig
+from .shard_config import ShardConfig
 from .sharder import ModelSharder
 from .shardformer import ShardFormer
 
-__all__ = ["ShardConfig", "ModelSharder", "ShardFormer", "PipelineGradientConfig"]
+__all__ = ["ShardConfig", "ModelSharder", "ShardFormer", "PipelineGradCkptConfig", "GradCkptCollection"]
